[P4ADEV-2313] Code review
