+++ src/main/java/it/gov/pagopa/payhub/activities/activity/exportflow/ExportFileExpirationHandlerActivity.java
@@ -4,12 +4,11 @@ import io.temporal.activity.ActivityInterface;
 import io.temporal.activity.ActivityMethod;
 
 /**
- * Interface for the ExportFileExpirationHandlerActivity.
  * Handles the expiration of an ExportFile identified by the provided ID.
  */
 @ActivityInterface
 public interface ExportFileExpirationHandlerActivity {
 
   @ActivityMethod
-  void handleExpiration(Long id, String codError);
+  void handleExpiration(Long exportFileId);
 }
