--- src/main/java/it/gov/pagopa/payhub/activities/exception/InvalidFileException.java
@@ -1,7 +1,0 @@
-package it.gov.pagopa.payhub.activities.exception;
-
-public class InvalidFileException extends RuntimeException {
-  public InvalidFileException(String message) {
-    super(message);
-  }
-}
